requirements.txt
flask
requests
pyjwt
flask_cors
